Update home.html
Hi mi friend, I downloaded your code to study it, I have to congratualte you for the splendid work. I found a small error in the script that executes the ajax call and returns the json. Precisely lines 93 and 94. Accessing the value with "data.fields.id" I discover that it is udefined. I solved it by directly accessing the "primary key" like this -> "data.pk". 
So it would be appropriate to change it like this: <a href="/update/'+data.pk+'" class="btn btn-success"> Update </a> and <a href = "/ delete / '+ data. pk + '"class =" btn btn-danger "> Delete </a>.  Thanks friend.